starting the revamp, now using an in memory db to cut down on API calls

+++ Osu_gene/user.py
@@ -19,7 +19,7 @@ from dotenv import load_dotenv
 load_dotenv()
 
 
-client_id = int(os.getenv('CLIENT_ID'))
+client_id = float(os.getenv('CLIENT_ID'))
 client_secret = os.getenv('CLIENT_SECRET')
 redirect_url = os.getenv('REDIRECT_URL')
